[proliferate] split cloud ORM model package
