+++ server/proliferate/db/models/cloud/__init__.py
@@ -1,0 +1,14 @@
+"""Cloud-domain ORM model package.
+
+Importing this package registers all cloud ORM tables with SQLAlchemy metadata.
+Callers should import concrete models from the owning module in this package.
+"""
+
+from . import credentials as credentials  # noqa: F401
+from . import mcp as mcp  # noqa: F401
+from . import mobility as mobility  # noqa: F401
+from . import repo_config as repo_config  # noqa: F401
+from . import runtime_environments as runtime_environments  # noqa: F401
+from . import sandboxes as sandboxes  # noqa: F401
+from . import workspaces as workspaces  # noqa: F401
+from . import worktree_policy as worktree_policy  # noqa: F401
