+++ server/proliferate/db/store/cloud_mcp/oauth_clients.py
@@ -6,7 +6,7 @@ from sqlalchemy import delete, select
 from sqlalchemy.ext.asyncio import AsyncSession
 
 from proliferate.db import engine as db_engine
-from proliferate.db.models.cloud import CloudMcpOAuthClient
+from proliferate.db.models.cloud.mcp import CloudMcpOAuthClient
 from proliferate.db.store.cloud_mcp.types import CloudMcpOAuthClientRecord
 from proliferate.utils.time import utcnow
