+++ server/proliferate/db/store/cloud_worktree_policy.py
@@ -11,7 +11,7 @@ from sqlalchemy.dialects.postgresql import insert as pg_insert
 from sqlalchemy.ext.asyncio import AsyncSession
 
 from proliferate.db import engine as db_engine
-from proliferate.db.models.cloud import CloudWorktreeRetentionPolicy
+from proliferate.db.models.cloud.worktree_policy import CloudWorktreeRetentionPolicy
 from proliferate.utils.time import utcnow
 
 DEFAULT_MAX_MATERIALIZED_WORKTREES_PER_REPO = 20
