+++ server/tests/unit/test_cloud_orm_models.py
@@ -1,0 +1,25 @@
+"""Cloud ORM package registration tests."""
+
+import proliferate.db.models.cloud  # noqa: F401
+from proliferate.db.models.base import Base
+
+
+def test_cloud_orm_package_registers_all_cloud_tables() -> None:
+    expected_tables = {
+        "cloud_runtime_environment",
+        "cloud_workspace",
+        "cloud_workspace_setup_run",
+        "cloud_workspace_mobility",
+        "cloud_workspace_handoff_op",
+        "cloud_sandbox",
+        "cloud_credential",
+        "cloud_worktree_retention_policy",
+        "cloud_mcp_connection",
+        "cloud_mcp_connection_auth",
+        "cloud_mcp_oauth_flow",
+        "cloud_mcp_oauth_client",
+        "cloud_repo_config",
+        "cloud_repo_file",
+    }
+
+    assert expected_tables <= set(Base.metadata.tables)
